Extracting Kapt/Java Workaround from main class
